+++ src/apsg/pandas/__init__.py
@@ -1,27 +1,17 @@
 # -*- coding: utf-8 -*-
 
 from apsg.pandas._pandas_api import (
-    VectorSetBaseAccessor,
-    APSGAccessor,
-    Vec3Accessor,
-    FolAccessor,
-    LinAccessor,
-    FaultAccessor,
-    Vector3Array,
+    FaultArray,
     FolArray,
     LinArray,
-    FaultArray,
+    Vector3Array,
+    pd,
 )
 
 __all__ = (
-    "VectorSetBaseAccessor",
-    "APSGAccessor",
-    "Vec3Accessor",
-    "FolAccessor",
-    "LinAccessor",
-    "FaultAccessor",
     "Vector3Array",
-    "FolArray",
     "LinArray",
+    "FolArray",
     "FaultArray",
+    "pd",
 )